WEBAPP-214: Remove PoiList and add PoiListItem
